fix(check): handle multiple tags
Since we look into an image tags to determine whether an
update is required or not, having multiple tags makes the
task harder as we don't know which one corresponds to the
remote version. In case an image has multiple tags, we
mark it as ambiguous and skip its treatment. To fix this
a container can be given an image reference as a label
and it will be used for all subsequent checks